+++ assets/src/edit-story/app/media/actions/useReloadMedia.js
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2020 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/**
+ * WordPress dependencies
+ */
+import { useCallback } from '@wordpress/element';
+
+function useReloadMedia( { setIsMediaLoading, setIsMediaLoaded } ) {
+	const reloadMedia = useCallback( () => {
+		setIsMediaLoading( false );
+		setIsMediaLoaded( false );
+	}, [ setIsMediaLoading, setIsMediaLoaded ] );
+
+	return reloadMedia;
+}
+
+export default useReloadMedia;
